fix: consolidate migrate tests
